Edit config classes
